Audio: Increase maximum allowed bitrate
Increase audio bitrate limit from 384 kbits/s to 2048 kbits/s.
This is to allow high fidelity audio for 5.1 and 7.1 surround audio,
which require higher bitrates.

The bitrate values are clamped, for codecs that would fail to render
with high bit rates set.

Also bitrate property is now not drawn for losless codecs, as it is not
used.

Pull Request: https://projects.blender.org/blender/blender/pulls/151126