Anim: Simplify function for relative shapekey evaluation
This refactors the relative shapekey evaluation function into a simpler,
more readable state.
Previously all shapekey evaluation happened in `key_evaluate_relative`
with very complex juggles of `char *` that are
actually `float *` and different strides and offsets.

It turns out *none* of the `char *` casting is required. No matter the
shapekey type, it is always a float3 and the shapekeys track indices into
a flat float array by assuming it is a float3 array.
I took advantage of that to massively simplify the function.

While I expect there to be no change in behavior, I did measure a change in performance.

Measurements for `key_evaluate_relative` in main and `key_evaluate_relative_float3` in PR

| main | PR |
| - | - |
| 3.64 ms | 2.44 ms |

Part of #149405

Pull Request: https://projects.blender.org/blender/blender/pulls/150710