Fix: Vulkan: Incorrect state transition
There is a case where the image layouts are administrated incorrectly. I
haven't found a case when this is actually an issue, but better to fix
it.

Subimage tracking may never alter the state of the global state of
the image.

Pull Request: https://projects.blender.org/blender/blender/pulls/151370