Geometry Nodes: Bone Info node for reading armature pose
New Bone Info node for exposing armature pose bone transforms in geometry nodes.

The node takes an armature object as input together with a bone name and outputs
various commonly used transforms of the bone as well as its rest length:
- Rest Pose: The orientation of the bone in the original armature, based only on
  the head and tail positions, without any additional pose mode transform.
- Pose: Final transform of the bone, after animation, drivers, and constraints
  are applied.
- Local Pose: Offset between the rest pose and final pose of a bone, both
  relative to the parent bone.
- Transform Pose: Offset between the rest pose and final pose of a bone, both
  relative to the parent bone, without  any constraints applied.

Currently, the patch adds a depsgraph dependency to the entire evaluated
armature instead of just the bone. This is because the bone name is not
generally known when building the depdendency graph and cannot be traced to a
single input string.

Co-authored-by:
- Zakmen150 <dimitri@screenskinz.co>
- Jacques Lucke <jacques@blender.org>
- Lukas Tönne <lukas@blender.org>
Pull Request: https://projects.blender.org/blender/blender/pulls/145870