FFMPEG: Add Option for Custom CRF Value
Add the option of manually inputting a custom CRF value for video
codecs that support CRF values.

This PR therefore introduces an additional Custom Quality option, which
allows the user to directly set the CRF value using a 0 to 1 value,
with 0 being the lowest possible video quality and 1 being lossless.
This not only allows the user to select values between Perceptually
Lossless and Lossless, it also makes it possible to fine tune the CRF
value once a satisfactory range has been found.

Motivation behind this PR:
With the currently available presets, large parts of the possible CRF
range are inaccessible. This is especially a problem with the gap
between Perceptually Lossless and Lossless. When uploading the rendered
video to basically any major video sharing platform, the platform
reencodes the already encoded video for better streaming delivery. This
means that effectively video quality is reduced twice, once when it's
encoded in Blender and once when it's encoded by the platform,
oftentimes leading to significant quality drops as a result. Ideally
you'd want to encode losslessly in Blender to let the platform have as
much information to work with as possible, however, the file sizes
resulting from lossless encoding are usually far too large to be
practical (~2GB per minute for 4K 60FPS when using H.265). On the other
hand, Perceptually Lossless, despite it's name, is in fact not
perceptually lossless as I've found in many cases. I know that the
FFMPEG website also claims it were, but at least at 4K resolution I can
clearly see quality reductions from Perceptually Lossless.
Therefore, for uploading videos online you'd usually want something
between Perceptually Lossless and Lossless. The file sizes should be
still somewhat reasonable, while the quality should be better than
Perceptually Lossless.