Refactor: Simplify shapkey copy function
No functional changes intended.

This simplifies the process of copying the shapekey data.

* `cp_cu_key` was removed. This was only ever called with the range `{0, tot}`
meaning it always copied all the nurbs segments anyway. Because of that all
vertices of either bpoint or beztriple were always copied. This was replaced
to a call to `copy_key_float3_range` which does not need to differentiate
between bpoint and beztriple.
* all calls to `copy_key` except for 1 were replaced with calls to `copy_key_float3_range`
* `copy_key` was converted into `copy_key_float3_weighted` to hint at what's
so special about this function. It was also greatly simplified.

Pull Request: https://projects.blender.org/blender/blender/pulls/151407