Refactor: Shapekey - remove range argument
No functional changes intended.

For curves, the range argument was passed into the copy and evaluate functions.
It seems like this was used in past blender versions because nurbs and bezier curves
needed to be handled differently.
However nowadays it is always a float3 array so the evaluate function does not need
to care about nurbs or bezier. The same applies to the copy function.

Since they are the same underlying data now, they can just be treated the same,
and evaluation/copy of a subset of the shapekey is no longer needed.

Pull Request: https://projects.blender.org/blender/blender/pulls/151415