Fix #151232: Sculpt blend brushes pick up color from masked vertices
Previously, when determining the "wet mix" color of the brush, the
sculpt paint brush would only use the distance and visibility to
determine if a given vertex should be sampled. This behaves
inconsistently with what people expect, so update it to use front face
filtering and automasking.

Pull Request: https://projects.blender.org/blender/blender/pulls/151347