Fix #151083: Separate by Material breaks custom normals part II
43d1685bc318 already fixed this for edit mode, but calling the operator
in object mode was still wrong.

When in object mode, ensure we have updated normals for the BMesh, so use
`BMeshFromMeshParams` for that.

NOTE: `BMeshFromMeshParams` `calc_face_normal` seems to enough to fix
this, I havent spent the time to track down in detail where this is
needed in the rabbit hole that normals calculation code is, other code
that relies in updated bmesh normals seems to do both `calc_face_normal`
**and** `calc_vert_normal` though...

Pull Request: https://projects.blender.org/blender/blender/pulls/151423