Fix #151039: EEVEE: Lookdev World doesn't rotate without blur
Contrary to the EEVEE legacy implementation, the original Background HDRI
is always used when the blur level is around 0.

The new implementation only rotated the probe data, thus the sharp background
was not correctly rotating.

To fix this we have to implement the rotation logic inside the internal lookdev
nodetree and disable this rotation setup when capturing the light probes which
have their fast rotation logic.

Pull Request: https://projects.blender.org/blender/blender/pulls/151386