UI: Default 'Save Modified Images' checkbox to enabled
Prior to this commit, the "Save Modified Images" checkbox when closing a
file would remember the previous value from when it was modified,
leading to unintuitive behavior (e.g. if the dialog box was
opened, modified and then the operation cancelled) and potential data
loss.

This commit resets the value when creating the dialog box.

Part of #134518

Pull Request: https://projects.blender.org/blender/blender/pulls/151450