Cleanup: Rename BLI_kdtree_op to BLI_kdtree
This is the last commit in a sequence of actions to move the
implementation of KDTree to templates. In order to do that, it was
necessary to replace the old main BLI header of kdtree with the
implementation file from intern, and to transfer the history of
implementation file and be able to build blender on any intermediate
commit, it was required to make 3 separate commit.

Pull Request: https://projects.blender.org/blender/blender/pulls/151338