Tests: Add unit tests for using bundled essential brushes
This commit adds a set of basic tests for all bundled brushes in
Sculpt, Vertex Paint, and Weight Paint. For each brush, we store the
initial state of the mesh, perform a stroke, and then compare the final
state of the mesh.

The attribute being compared is dependent on the mode and brush being
used. In general we verify that the attribute has changed since the
initial state, and that it is valid.

This does not yet enumerate each of the special modes for some brushes
nor does it handle certain classes of brushes (sculpt simulation) or
brushes that may have more specific requirements for positioning.

This currently only works on a mesh, though more dimensions (e.g.
multires, modifiers, dyntopo) should be added.

This is intended to work in tandem with the general render tests:
these unit tests can enumerate many possible permutations, whereas
the render tests can verify the appearance of a deformation.

Pull Request: https://projects.blender.org/blender/blender/pulls/151336