Fix: VSE thumbnail generation could deadlock
Previously, there was a chance that when caluclating the next thumbnail
timeline postion, we would get the exact same position.

This would lead to a dead lock as the thumbnail generation would never
advance and get to the end of the while loop.

Now we instead only try to do the non safe incremental calulation once.

Pull Request: https://projects.blender.org/blender/blender/pulls/151672