Refactor: Cycles: Create PointerRNA properly with owner_id
This technically works without the owner_id, but it shouldn't
be done and may break in the future. Unfortunately this involves
passing `bScreen` nearly everywhere alongside the 3D view, to
construct `BlenderViewportParameters`. The `bScreen` pointer
is null when the 3D view pointer is null.

Pull Request: https://projects.blender.org/blender/blender/pulls/151791