GPU: Shader: Ensure types used in buffers are checked by shader tool
This adds a alias macro to all struct parsed with the
`[[host_shared]]` attribute. This macro is then used inside
shared struct declaration to make sure the type was indeed linted.
This also ensures host_shared structures only contain
structures that are also host_shared.

In some cases, the structure type is not aligned to 16bytes
but is still valid (e.g. `Position`). This is not yet supported
by the `host_shared` linter. For these cases, we disable
the use of the `linted` aliases when declaring buffers.
This is also the case for python shaders where the layouts
can contains macros and we cannot lint them.

We also add the same `[[host_shared]]` attribute to enum types
to make sure they are the right size when included inside
buffer structs. This also means that non 32bits enums are now
possible in shared headers as long as they are not used in
shared structs.

Pull Request: https://projects.blender.org/blender/blender/pulls/151798