Python: introduce Disk File Hash Service
Introduce a way to efficiently hash files, and to check whether a file
on disk matches an expected hash. "Efficient" here means that the
hashes are cached, and the cached value is returned when the file on
disk still has the same size & last-modification timestamp.

This was made for the [Remote Asset Libraries project][1]. In the
future, this system will also be used for the [BAT v2 project][2], as
that also will need to compute hashes for files.

[1]: https://projects.blender.org/blender/blender/issues/134495 [2]:
https://projects.blender.org/studio/flamenco/issues/104437

This PR introduces the following classes:

- `DiskFileHashService`: The top-level 'service' instance that's meant
  to be used by other code. Takes care of computing hashes for files.
  It can cache these hashes, and check them for 'freshness' based on
  the file's size in bytes and last-modified timestamp.
- `DiskFileHashBackend`: Protocol class that defines the API for file
  hash storage backends.
- `SQLiteBackend`: Implementation of that protocol, using SQLite as a
  way to store hashes.

More info in the PR.

Pull Request: https://projects.blender.org/blender/blender/pulls/151497