DNA: support typed ListBase
Since #134531 is merged now, it's possible to use C++ types in DNA within
limitations. The most prominent case where we benefit from this is to replace
`ListBase` with a templated type. This improves type-safety and makes iterating
over the list easier.

This works by adding a new `ListBaseT<Type>` template which is a thin wrapper
around the existing `ListBase` type. In DNA, `ListBaseT` can be used as a
drop-in replacement for `ListBase`. This does not affect file-compatibility,
because it's still written as old-style `ListBase` in .blend files.

The DNA parsing code has been updated to ignore suffixes of type names of the
form `T<...>`. This results in `ListBaseT<...>` as being parsed as just
`ListBase`. This could be generalized further but it does not seem obviously
worth it right now.

For testing, I used the new `ListBaseT` for nodes, links and sockets. It can be
used for all the other `ListBase` members in DNA separately.

Pull Request: https://projects.blender.org/blender/blender/pulls/138480