Refactor: BLO: use new BlendWriter::write_struct method
This is part of #151842.

For the most part the changes here were done automatically, but it was a bit
more involved than originally planned. That's because the pointer type that was
passed to `BLO_write_struct` did not align with the passed in type name in a few
cases. So just doing the proposed global replacement could easily lead to bugs
where the wrong type would be written.

To solve this, I did the following:
1. Do a global replacement of this form:
   ```cpp
   // Old
   BLO_write_struct(writer, bNodeSocket, sock);
   // New
   writer->write_struct<bNodeSocket>(sock);
   ```
2. This allowed me to use compiler errors to find all the cases where the
   pointers did not match. Those cases I fixed by adding a separate
   `BlendWriter::write_struct_cast` method which expects the type name to be
   passed in explicitly, instead of being deduced from the pointer.
3. Remove the redundant template argument so that most lines end up like so:
   `writer->write_struct(sock)`.

Ideally, `write_struct` could reliably check for that the type is right, but I
didn't find a way to do that yet. My main attempt was to compare `sizeof(T)`
with `MEM_allocN_len(data)` but that didn't work because sometimes the pointers
passed to `write_struct` are not allocated with the guarded allocator (e.g. when
a local copy is made just for writing). My hope is that in new code, these kinds
of bugs will result in obviously wrong behavior when implementing new features.
It's not that often that we have dynamically typed pointers fortunately.

Pull Request: https://projects.blender.org/blender/blender/pulls/151975