Introduce BVH export option to sort bone child node name appearances
When an armature object is exported into the BVH format, the order of
appearance of each bone's child nodes is determined by the order in which
they were originally created or imported. This can be a problem for
downstream asset consumption by technical animators for tasks like
retargeting in certain project pipelines because the structure of the motion
data expressed by a BVH file is dependent on the ordering of the
transformation delta channels as specified by those corresponding joints.

This commit introduces a new checkbox option into the menu for exporting
with the BVH format which will allow any user to decide whether to sort
bone children by their node names alphabetically. In addition to addressing
what has been described above, if other tools are in use besides Blender
in such an environment, this feature can allow for multiple comparable
skeletons to have their hierarchy expression order unified through Blender
across a variety of different sets of motion data. Yet another resolution is
detailed in a forum post below...
https://blenderartists.org/t/how-does-one-reorder-child-bones-within-an-armature/684994/3

The present solution is limited in that it does not provide any greater
customization of the sorting option, but rather only supports sorting
alphabetically.

Co-authored-by: Pierce Brooks <piercebrooks@rocketmail.com>
Pull Request: https://projects.blender.org/blender/blender/pulls/151626