Fix: Vulkan: Validation error when using conservative depth testing
Conservative depth prepass of the overlay engine can read an additional
2 bytes of the index buffer when there are an uneven number of indices
and the indices are stored as 16 bit.

This PR fixes the validation error by using the actual allocated size
and not the used size of the index buffer to allow perfectly safe
out of bounds read.

Pull Request: https://projects.blender.org/blender/blender/pulls/152028