UI: Quad View Simplification
This PR makes a number of changes to the 3D Viewport when in "Quad
View" mode in an attempt to simplify it a bit. The navigate gizmo,
if enabled, is only shown in the top-right quadrant. The General
Info, Performance, and Statistics are only shown in the top-left
quadrant. Except for the view name, still shown in all quadrants, but
with simplified names (without "Orthographic") in the locked quadrants.

Pull Request: https://projects.blender.org/blender/blender/pulls/152090