Fix: USD: Handle displayColor in more situations for Shape primitives
If `displayColor` was used with an interpolation type other than
'vertex', we would attempt to apply the attribute to the faces of the
mesh before we actually built those faces.

Read in the color information once the mesh has been fully built.

This would assert in debug builds and write to invalid memory locations
in retail.

Pull Request: https://projects.blender.org/blender/blender/pulls/152141